Updated qwen2.5vl
