+++ mlx_vlm/models/qwen2_5_vl/language.py
@@ -8,7 +8,24 @@ import numpy as np
 
 from ..base import LanguageModelOutput, create_attention_mask
 from ..cache import KVCache
+from .vision import VisionConfig
 
+@dataclass
+class ModelConfig:
+    image_token_index: int 
+    video_token_index: int 
+    vision_start_token_id: int 
+    vision_config: VisionConfig
+
+    @classmethod
+    def from_dict(cls, params):
+        return cls(
+            **{
+                k: v
+                for k, v in params.items()
+                if k in inspect.signature(cls).parameters
+            }
+        )
 
 @dataclass
 class TextConfig:
@@ -273,11 +290,13 @@ class Qwen2Model(nn.Module):
 
 
 class LanguageModel(nn.Module):
-    def __init__(self, args: TextConfig):
+    def __init__(self, args: TextConfig, config: ModelConfig):
         super().__init__()
         self.args = args
+        self.config = config
         self.model_type = args.model_type
         self.model = Qwen2Model(args)
+        self.rope_deltas = None
 
         if args.model_type != "qwen2_5_vl":
             raise ValueError(f"Unsupported model type: {args.model_type}")
@@ -285,14 +304,163 @@ class LanguageModel(nn.Module):
         if not args.tie_word_embeddings:
             self.lm_head = nn.Linear(args.hidden_size, args.vocab_size, bias=False)
 
+    def get_rope_index(
+        self,
+        input_ids: mx.array,
+        image_grid_thw: Optional[mx.array] = None,
+        video_grid_thw: Optional[mx.array] = None,
+        attention_mask: Optional[mx.array] = None,
+    ):
+        # Calculate RoPE index for image/video tokens
+        batch_size, seq_length = input_ids.shape
+        position_ids = mx.arange(seq_length, dtype=mx.int32)
+        position_ids = mx.broadcast_to(position_ids[None, :], (batch_size, seq_length))        
+        spatial_merge_size = self.config.vision_config.spatial_merge_size
+        image_token_id = self.config.image_token_index
+        video_token_id = self.config.video_token_index
+        vision_start_token_id = self.config.vision_start_token_id
+        mrope_position_deltas = []
+        if input_ids is not None and (image_grid_thw is not None or video_grid_thw is not None):
+            total_input_ids = input_ids
+            if attention_mask is None:
+                attention_mask = mx.ones_like(input_ids)
+            position_ids = mx.ones(
+                (3, input_ids.shape[0], input_ids.shape[1]), dtype=input_ids.dtype
+            )
+            image_index, video_index = 0, 0
+            for i, input_ids in enumerate(total_input_ids):
+                input_ids = mx.where(attention_mask[i] == 1, input_ids, mx.zeros_like(input_ids))
+                image_nums, video_nums = 0, 0
+                vision_start_indices = mx.sum(mx.where(input_ids == vision_start_token_id, mx.arange(input_ids.shape[0]), mx.zeros_like(input_ids)))
+                vision_tokens = input_ids[vision_start_indices + 1]
+                image_nums = (vision_tokens == image_token_id).sum().item()
+                video_nums = (vision_tokens == video_token_id).sum().item()
+                input_tokens = input_ids.tolist()
+                llm_pos_ids_list: list = []
+                st = 0
+                remain_images, remain_videos =  image_nums, video_nums
+                for _ in range(image_nums + video_nums):
+                    if image_token_id in input_tokens and remain_images > 0:
+                        ed_image = input_tokens.index(image_token_id, st)
+                    else:
+                        ed_image = len(input_tokens) +1
+                    if video_token_id in input_tokens and remain_videos > 0:
+                        ed_video = input_tokens.index(video_token_id, st)
+                    else:
+                        ed_video = len(input_tokens) +1
+                    if ed_image < ed_video:
+                        t, h, w = (image_grid_thw[image_index][0], image_grid_thw[image_index][1], image_grid_thw[image_index][2])
+                        image_index += 1
+                        remain_images -= 1
+                        ed = ed_image
+                    else:
+                        t, h, w = (video_grid_thw[video_index][0], video_grid_thw[video_index][1], video_grid_thw[video_index][2])
+                        video_index += 1
+                        remain_videos -= 1
+                        ed = ed_video
+                    llm_grid_t , llm_grid_h, llm_grid_w = (t.item(), h.item() // spatial_merge_size, w.item() // spatial_merge_size)
+                    text_len = ed - st
+                    st_idx = llm_pos_ids_list[-1].max() + 1 if len(llm_pos_ids_list) > 0 else 0
+                    index = mx.arange(text_len).reshape(1, text_len)
+                    index = mx.broadcast_to(index, (3, text_len))
+                    index = index + st_idx
+                    llm_pos_ids_list.append(index)
+                    t_index = mx.arange(llm_grid_t).reshape(llm_grid_t, 1)  # Equivalent to .view(-1, 1)
+                    t_index = mx.broadcast_to(t_index, (llm_grid_t, llm_grid_h * llm_grid_w))  # Equivalent to expand()
+                    t_index = t_index.flatten()  # Flattens to 1D
+
+                    h_index = mx.arange(llm_grid_h).reshape(1, llm_grid_h, 1)  # Equivalent to .view(1, -1)
+                    h_index = mx.broadcast_to(h_index, (llm_grid_t, llm_grid_h, llm_grid_w))  # Equivalent to expand()
+                    h_index = h_index.flatten()  # Flattens to 1D
+
+                    w_index = mx.arange(llm_grid_w).reshape(1, 1, llm_grid_w)  # Equivalent to .view(1, -1)
+                    w_index = mx.broadcast_to(w_index, (llm_grid_t, llm_grid_h, llm_grid_w))  # Equivalent to expand()
+                    w_index = w_index.flatten()  # Flattens to 1D
+
+
+                    llm_pos_ids_list.append(mx.stack([t_index, h_index, w_index]) + text_len + st_idx)
+                    st = ed + llm_grid_t * llm_grid_h * llm_grid_w
+                if st < len(input_tokens):
+                    st_idx = llm_pos_ids_list[-1].max() + 1 if len(llm_pos_ids_list) > 0 else 0
+                    text_len = len(input_tokens) - st
+
+                    t_index = mx.arange(text_len).reshape(1, text_len)  # Equivalent to .view(-1, 1)
+                    t_index = mx.broadcast_to(t_index, (3, text_len))  # Equivalent to expand(3, -1)
+
+                    llm_pos_ids_list.append(t_index + st_idx)
+
+                llm_positions = mx.concatenate(llm_pos_ids_list, axis=1).reshape(3, -1)
+                mask = mx.array(attention_mask[i] == 1)
+                expanded_mask = mx.expand_dims(mask, axis=0)
+                expanded_mask = mx.broadcast_to(expanded_mask, (3,1, mask.shape[0]))
+                expanded_positions = mx.expand_dims(llm_positions, axis=1)
+                new_positions = mx.where(expanded_mask, expanded_positions, position_ids[:, i:i+1, :])
+                updated_position_ids = mx.concatenate([
+                    position_ids[:, :i, :],
+                    new_positions,
+                    position_ids[:, i+1:, :]
+                ], axis=1)
+                position_ids = updated_position_ids
+                mrope_position_deltas.append(llm_positions.max() + 1 - len(total_input_ids[i]))
+            mrope_position_deltas = mx.array(mrope_position_deltas)[0]
+            return position_ids, mrope_position_deltas
+        else:
+            if attention_mask is not None:
+                position_ids = mx.cumsum(attention_mask.astype(mx.int64), axis= -1) - 1
+                position_ids = mx.where(attention_mask == 0, mx.ones_like(position_ids), position_ids)
+                position_ids = mx.expand_dims(position_ids[0], axis=0)
+                position_ids = mx.tile(position_ids, (3, 1, 1))
+                max_position_ids = position_ids.max(0, keepdims=False)[0].max(-1, keepdims=True)[0]
+                mrope_position_deltas = max_position_ids + 1 - attention_mask.shape[-1]
+            else:    
+                position_ids = mx.arange(input_ids.shape[1]).reshape(1, -1)
+                position_ids = mx.broadcast_to(position_ids, (3, input_ids.shape[0], input_ids.shape[1]))
+                mrope_position_deltas = mx.zeros(
+                    [input_ids.shape[0], 1],
+                    dtype=input_ids.dtype,
+                )
+            return position_ids, mrope_position_deltas
+        
     def __call__(
         self,
         inputs: mx.array,
         inputs_embeds: Optional[mx.array] = None,
         mask: Optional[mx.array] = None,
         cache=None,
-        position_ids: Optional[mx.array] = None,
+        **kwargs,
     ):
+        position_ids = kwargs.pop("position_ids", None)
+        pixel_values = kwargs.pop("pixel_values", None)
+        image_grid_thw = kwargs.pop("image_grid_thw", None)
+        video_grid_thw = kwargs.pop("video_grid_thw", None)
+        input_ids = kwargs.pop("input_ids", inputs)
+        # reset rope_deltas when processing a new image/video
+        if pixel_values is not None:
+            self.rope_deltas = None
+
+        if position_ids is None and (mask is None or mask.ndim == 2):
+            # Calculate RoPE index once per generation in the pre-fill stage only
+            if (
+                (cache is not None and cache[0] == 0)
+                or self.rope_deltas is None
+                or cache is None
+            ):
+                position_ids, rope_deltas = self.get_rope_index(
+                    input_ids, image_grid_thw, video_grid_thw, mask
+                )
+                self.rope_deltas = rope_deltas
+            else:
+                # Use the prev pre-calculated rope-deltas to get the correct position ids
+                batch_size, seq_length = input_ids.shape
+                delta = cache[-1].offset + self.rope_deltas if cache is not None else 0
+                delta = delta[None][None]
+                position_ids = mx.arange(seq_length).reshape(1, seq_length)
+                position_ids = mx.broadcast_to(position_ids, (batch_size, seq_length))
+                if cache is not None:
+                    # Repeat delta for each batch
+                    delta = mx.repeat(delta, batch_size // delta.shape[0], axis=0)
+                position_ids = mx.add(position_ids, delta).reshape(position_ids.shape)
+                position_ids = mx.broadcast_to(position_ids, (3, batch_size, seq_length))
         out = self.model(inputs, mask=mask, cache=cache, inputs_embeds=inputs_embeds, position_ids=position_ids)
         if self.args.tie_word_embeddings:
             out = self.model.embed_tokens.as_linear(out)
